% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lperd_mtm.R
\name{lperd.mtm}
\alias{lperd.mtm}
\title{Multitaper Log-Periodogram}
\usage{
lperd.mtm(x,nw,k)
}
\arguments{
\item{x}{\emph{N} by \emph{n} matrix containing \emph{n} training time series each with length \emph{N}.}

\item{nw}{Width of tapers used in multitaper spectral estimation.}

\item{k}{Number of tapers used in multitaper spectral estimation.}
}
\value{
\item{freq}{Fourier frequencies from 0 to 1.}
\item{lspec}{Log-periodogram.}
\item{spec}{Periodigram on the natural scale.}
}
\description{
Obtain multivariate log-periodogram.  Not often used by itself, but is used as part of \code{cep.mtm}.
}
\author{
Robert Krafty \email{rkrafty@pitt.edu}
}
\references{
Krafty, RT (2016) Discriminant Analysis of Time Series in the Presence of Within-Group Spectral Variability.  \emph{Journal of Time series analysis}
}
\seealso{
\code{\link{cep.mtm}}, \code{\link{cep.lda}}
}

