% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cep_lda.R
\name{plot.ceplda}
\alias{plot.ceplda}
\title{Plot Method for Class 'ceplda'}
\usage{
\method{plot}{ceplda}(x, ...)
}
\arguments{
\item{x}{an object of class "ceplda".}

\item{...}{additional arguments.}
}
\description{
Plots a set of data on one, two or more linear discriminants.
}
\details{
This function is a method for the generic function \code{plot()} for class "\code{ceplda}". It can be invoked by calling \code{plot(x)}
for an object \code{x} of the appropriate class, or directly by calling \code{plot.ceplda(x)} regardless of the class of the object. Details usage of this function
is equivalent to \code{plot.lda(MASS)}.
}
\seealso{
\code{\link{cep.lda}}, \code{\link{plot.lda}}
}

