% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cep_lda.R
\name{print.ceplda}
\alias{print.ceplda}
\title{Print Method for Class 'ceplda'}
\usage{
\method{print}{ceplda}(x, ...)
}
\arguments{
\item{x}{an object of class "ceplda".}

\item{...}{additional arguments.}
}
\description{
Print the results from \code{cep.lda}
}
\seealso{
\code{\link{cep.lda}}
}

